Renamed bullet.rs => ammo.rs
