+++ libplen/src/ammo.rs
